feat: Add final chart interaction enhancements
This commit delivers the final set of enhancements to the web application's chart, focusing on intuitive navigation and better UI control.

Key Features:
- **Default Pan and Zoom:** The chart's default behavior is now set to pan when clicking and dragging, while the mouse wheel continues to control zooming. This provides a more fluid and TradingView-like user experience without needing to select tools from a toolbar.
- **Individual Symbol Removal:** Users can now remove individual stocks from the comparison view by clicking an 'x' button on the symbol's tag. The chart automatically updates to reflect the change.
- **UI and Layout Fixes:** The CSS layout has been refactored to use Flexbox, resolving click interception issues and ensuring the UI is robust.
